+++ pysys-examples/internal/testcases/PySys_internal_113/Input/subdir1/PySys_cor_001/run.py
@@ -1,0 +1,11 @@
+from pysys.constants import *
+from pysys.basetest import BaseTest
+from pysys.exceptions import *
+
+class PySysTest(BaseTest):
+	def execute(self):
+		self.log.info('This is a normal PySys test')
+		self.addOutcome(PASSED)
+
+	def validate(self):
+		pass